+++ src/pages/Autolending/AutolendingMessagingPage.vue
@@ -1,12 +1,6 @@
 <template>
-	<www-page :gray-background="true">
-		<div class="title-area">
-			<div class="row column">
-				<h1>Auto-lending</h1>
-				<h2>Make the impact you want even if you’re away from your account for a while</h2>
-			</div>
-		</div>
-		<div class="row autolending-opt-out-content">
+	<div class="row autolending-messaging-content">
+		<template v-if="action == 'opt-out'">
 			<div data-test="opt-out-success" class="small-12 medium-10 columns" v-if="success">
 				<!-- Success -->
 				<p>You’re successfully opted out of auto-lending.</p>
@@ -25,26 +19,39 @@
 					<a href="mailto:contactus@kiva.org?subject=Auto-lending%20opt-out%20page">contactus@kiva.org</a>
 				</p>
 			</div>
-		</div>
-	</www-page>
+		</template>
+		<template v-if="action == 'disable'">
+			<div data-test="disable-success" class="small-12 medium-10 columns" v-if="success">
+				<!-- Success -->
+				<p>You successfully turned off auto-lending.</p>
+				<!-- eslint-disable-next-line -->
+				<p>Want to turn it back on or customize your settings? Just click <router-link to="/settings/autolending">here</router-link>.</p>
+			</div>
+
+			<div data-test="disable-failure" class="small-12 medium-10 columns" v-if="!success">
+				<!-- Failure -->
+				<p>Oops - this link no longer works.</p>
+				<p>
+					If you want to turn off auto-lending, please let us know at
+					<a href="mailto:contactus@kiva.org?subject=Auto-lending%20disable%20page">contactus@kiva.org</a>
+				</p>
+			</div>
+		</template>
+	</div>
 </template>
 
 <script>
-import WwwPage from '@/components/WwwFrame/WwwPage';
 
 export default {
 	props: {
 		success: {
 			type: Boolean,
 			default: false
+		},
+		action: {
+			type: String,
+			default: ''
 		}
-	},
-	components: {
-		WwwPage,
-	},
-	data() {
-		return {
-		};
 	},
 };
 </script>
@@ -54,11 +61,5 @@ export default {
 
 .autolending-opt-out-content {
 	padding: 2rem 0 6rem;
-}
-
-.title-area {
-	padding: 1.625rem 0;
-	margin-bottom: 2rem;
-	background-color: $white;
 }
 </style>
